+++ aquadoggo/src/graphql/client/query.rs
@@ -10,10 +10,10 @@ use crate::graphql::client::{EntryArgsRequest, EntryArgsResponse};
 
 /// The GraphQL root for the client api that p2panda clients can use to connect to a node.
 #[derive(Default, Debug, Copy, Clone)]
-pub struct Query;
+pub struct ClientRoot;
 
 #[Object]
-impl Query {
+impl ClientRoot {
     /// Return required arguments for publishing the next entry.
     async fn next_entry_args(
         &self,
@@ -50,17 +50,18 @@ mod tests {
     use async_graphql::Response;
     use p2panda_rs::entry::{LogId, SeqNum};
     use serde_json::json;
+    use tokio::sync::broadcast;
 
-    use crate::config::Configuration;
-    use crate::context::Context;
     use crate::graphql::client::EntryArgsResponse;
-    use crate::server::build_server;
+    use crate::http::build_server;
+    use crate::http::HttpServiceContext;
     use crate::test_helpers::{initialize_store, TestClient};
 
     #[tokio::test]
     async fn next_entry_args_valid_query() {
+        let (tx, _) = broadcast::channel(16);
         let store = initialize_store().await;
-        let context = Context::new(store, Configuration::default());
+        let context = HttpServiceContext::new(store, tx);
         let client = TestClient::new(build_server(context));
 
         // Selected fields need to be alphabetically sorted because that's what the `json` macro
@@ -103,8 +104,9 @@ mod tests {
 
     #[tokio::test]
     async fn next_entry_args_error_response() {
+        let (tx, _) = broadcast::channel(16);
         let store = initialize_store().await;
-        let context = Context::new(store, Configuration::default());
+        let context = HttpServiceContext::new(store, tx);
         let client = TestClient::new(build_server(context));
 
         // Selected fields need to be alphabetically sorted because that's what the `json` macro