+++ aquadoggo/src/lib.rs
@@ -18,13 +18,12 @@ mod config;
 mod context;
 mod db;
 mod errors;
-/// Aquadoggo graphql types for handling client and replication requests
 pub mod graphql;
+mod http;
 mod manager;
 mod materializer;
 mod node;
 mod replication;
-mod server;
 
 #[cfg(test)]
 mod test_helpers;