+++ CHANGELOG.md
@@ -21,7 +21,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - GraphQL endpoint for publishing entries [#123](https://github.com/p2panda/aquadoggo/pull/132)
 - Implement SQL `DocumentStore` [#118](https://github.com/p2panda/aquadoggo/pull/118)
 - Implement SQL `SchemaStore` [#130](https://github.com/p2panda/aquadoggo/pull/130)
+- Reduce and dependency tasks [#144](https://github.com/p2panda/aquadoggo/pull/144)
 - GraphQL endpoints for replication [#100](https://github.com/p2panda/aquadoggo/pull/100)
+- Inform materialization service about new operations [#161](https://github.com/p2panda/aquadoggo/pull/161)
 
 ### Changed
 
@@ -34,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Move `SqlStorage` into shared `Context` [#135](https://github.com/p2panda/aquadoggo/pull/135)
 - Refactor tests to use fixtures exported from `p2panda-rs` [#147](https://github.com/p2panda/aquadoggo/pull/147)
 - Use `DocumentViewId` for previous operations [#147](https://github.com/p2panda/aquadoggo/pull/147)
+- Use `VerifiedOperation` [#158](https://github.com/p2panda/aquadoggo/pull/158)
 
 ### Fixed